Move LibraryID from Thread_Utils to Common namespa
This commit moves the LibraryID struct into the Common namespace.
(This is to allow other namespaces to use it.)

This commit also removes the Common::Thread_Utils::Get_Library_Name()
function. (It's redundant, if you have the LibraryID struct (which is
needed for the function to work) then you can access the Name
pointer directly.)