Fixes for FileUtills::CopyPath().
This commit makes FileUtills::CopyPath() use streamsize instead of
size_t for file offsets.

This commit also makes FileUtills::CopyPath() return new error codes for
when the call to FileUtills::IsFileOrDirectory() fails.

This commit also changes some of FileUtills::CopyPath() error codes to
be more consistant.

(error code -7 is now -5 (Could not get parent directory.),
error code -8 is now -6. (Could not copy all data.))

As a result of this commit the bug of a directory being sent to
FileUtills::CopyFile() has been corrected.

(Previously the result from FileUtills::IsFileOrDirectory() was not
handled correctly, causing the directory to be given to FileUtills::CopyFile().)