Make FileUtills::CopyFile() aware of directories.
This commit makes FileUtills::CopyFile() aware of directories. (I.e. it
will return an error code if src or dest is a directory.)

The following error codes have been added to FileUtills::CopyFile():

	- Returns -17 if the given source file was a directory.
        - Returns -27 if the given dest file was a directory.
	- Returns -33 if FileUtills::IsFileOrDirectory() returns -3. (OS /
Arch not supported.)
	- Returns -34 if FileUtills::IsFileOrDirectory() returns -4. (A
permissions error occured.)
	- Returns -35 if FileUtills::IsFileOrDirectory() returns -5. (The
given path is empty.)
	- Returns -36 if FileUtills::IsFileOrDirectory() returns -6. (A
path componet does not exist.)
	- Returns -37 if FileUtills::IsFileOrDirectory() returns -7. (The
path has a file in it and is not at the end. (I.e you are treating a
file as a direct
	- Returns -39 if FileUtills::IsFileOrDirectory() returns -9. (All
other errors.)