Add Fatal Error Handler, rewrite Error Handler as.
This commit adds a Fatal Error Notification Subsystem implementation
and rewrites the pre-existing error handler code to be C code with C++
bindings.

Do note that the transition to C code means that an API change cannot be
avoided. Although attempts have been made to make the transition as
easy as possible.

Also the error table API version numbers have been updated. In addition
to adding an API call to check what those version numbers are.