Inital rewrite commit.
This commit is all over the place as it contains a lot of rewriten
functions, new API calls, code path restructuring, and even some new
error codes.

In addition most of this code was written over the course of the last
year, but was not completed, due to other work. Probably a benifit as
the original rewrite was "finalized" as far as it's code path
restructuring and API calls were concerned, but looking at it now, there
is some other modifications that need to be made that I was not
considering at the time.

Currently work is being done to port the code to C, as well as fix the
API calls to make all functions return Common error codes. (Any other
data that needs to be returned, will be moved to the arguments list as
pointers.) We are also elimiating the Common::commonLastErrorCode global
variable to make the API thread safe. In addition to makeing the API
more thread safe in general.