Fix FileUtills::ResolveSystemSymoblicLink_Syscall
This commit makes the "Devil's loop" in the
FileUtills::ResolveSystemSymoblicLink_Syscall() actually resolve
misbehaving symbolic links that are not /proc/self/exe. (Rather than
resolving /proc/self/exe which was not what was intended. Up the score
for copy-pastea.)

Also this commit makes FileUtills::ResolveSystemSymoblicLink_Syscall()
conform to the new FileUtills API requirements.