Commit missing file.
This commit adds the missing Dynamic_Library_Subsystem_Windows.c file.
(It's defined to be built in master, but not present.

We are creating this branch as it requires the fake stdbool header from
the error_handler branch to build under MSVC. Also due to the code being
untested, we are moving it here to avoid messing up master.)