Rename Panic::ERROR
This commit renames class Panic::ERROR to Panic::Panic_Error.

(Rant: Normally I would not permit this, but Panic is a Core API that
is rarely used. It was easier to change it than it was to fix the Windows
header files. (Apperently they define a ERROR of some kind that, somehow
despite ours being in our own Panic namespace, overrides ours. (Actually
the Windows headers undef it for some reason. Giving us weird compiler errors
like untagged classes.)) Whoever at Microsoft that thought "forbidding the use
of the identifier ERROR from a different namespace is a great idea!" should be slapped.....
/Rant)