Commit missing FileUtills_Private_API_Windows_Syscall.cpp
This commit adds the missing FileUtills_Private_API_Windows_Syscall.cpp
file.

Note: Currently this is just a stub file based on FileUtills_Private_API_Posix_Syscall.cpp
to allow CMake to finish generating it's makefile. It will not build on a
Windows system due to not actually implementing the needed functions properly.
(Well minus one that was not gutted prior to commit that is actually the Posix
function, but that is beside the point.)