Create DataProcess.c
This commit creates a seperate file (DataProcess.c) for the DataProcess code that can
be used by a C program.

The point of this is to allow the use of some DataProcess functionality by C programs
where possible.

To kickoff the new DataProcess.c file, this commit also moves the DataProces::Trivial_Random_Number_Generator() function code
to the new C file and renames it to DataProcess_Trivial_Random_Number_Generator().

The original DataProcess::Trivial_Random_Number_Generator() function now acts as a wrapper
to the DataProcess_Trivial_Random_Number_Generator() C function.