Use reassignment of given pointer for FileUtills_Get_Last_Path_Component()
This commit changes the prototype of FileUtills_Get_Last_Path_Component() to eliminate the first two arguments in favor of using a reassignable pointer and changable length argument set.
(I.e. Remove the constant vars, and reuse the given pointer and lengths to store the results if successfull. Code will NOT deallocate the original pointer, and assumes the user has made a copy of that pointer if it is still needed prior to calling the function.)

The purpose of this modification is to allow for easier integration with other functions such that a copy of the given c-string is not needed.
(For example if the function was called by another engine function that was already processing a working copy. This elimenates the need to make another copy of the c-string prior to calling this function.)