Fix issue with random seed generation in DataProcess_Trivial_Random_Number_Generator()
This commit fixes a random seed generation issue if DataProcess_Trivial_Random_Number_Generator() is called more than once (with reset_rand set) in the timeframe of one second.
(The issue was the C standard only supports a minimum resolution of one second in it's time functions.
If DataProcess_Trivial_Random_Number_Generator() had to reset it's random seed more than once within a second, it would get the same value from time() for the random seed.
This commit fixes that by using a static variable within the function to modifiy the returned time value so that the seed will be different from the previous one.)