Clean up DataProcess_Reallocate_C_String_With_NULL_Terminator().
This commit simplifies the check for whether or not a byte should be added to the given string.
(By altering the unary and removing the if block above it.)
Also makes the function always do the check to see if the reallocated string is NULL byte
terminated, instead of only when a string length change did not occur.