Condense DataProcess_Get_SubString_Using_Delimiter().
This commit removes repeated code that was used to handle the different codepaths that could be taken depending on getPriorData's value in DataProcess_Get_SubString_Using_Delimiter().
(I.e. It's repeated code, with only a few differences in checked values and arugments to memcpy. This redundant code has been removed.)