Add member variable for storing path length to Common_Dynamic_Library_Subsystem_Loaded_Dynamic_Library_Private
This commit adds pathToLibraryLength as a member variable to Common_Dynamic_Library_Subsystem_Loaded_Dynamic_Library_Private to keep track of the length of the pathToLibrary string in bytes.
This commit also modifies the required functions to store and return this new information in both the public and private API source files.