Commit missing Renderer_Subsystem....
OK, I thought I commited this over year ago, but I guess I didn't.

Unfortuntely the only remaining copy of this that I had was on a failed harddrive and the git subdirectory was irretrievable. As such this code is being recommited as is.

This code is C++ code, in addition a small set of Unit_Tests exist for it. I'm planning on converting this code over to C for easier use.