Partial compile fix for FileUtills_Private_API_Windows_Syscall.cpp. (Error log channels)
This commit is a partial compile fix for FileUtills_Private_API_Windows_Syscall.cpp due to the addition of error log channel support in Common_Error_Handler.

Note: The reason this commit is a partial, is because it only fixes the logging function calls nothing else. (This file is currently undergoing a rewrite and cannot be used currently.)