Workaround CMake link library error.
This commit is a workaround for linking the Common library.

Note: Common really should be staticly linked to Core (or at least the Core memory management functions...) but when building a shared library in CMake there is no built in support for doing this. Attempting to do so with GCC as the compiler will result in a "recompile with -fPIC" error. So as a temperary workaround, we are setting the Common shared library to link against the Core shared library. Until I find a way around this....